# rest/api/v1/openapi.yml
...
openapi: 3.1.0
paths:
...
  /api/v1/blocks/{hashOrNumber}:
    get:
      summary: Get block by hash or number
      description: Returns the block information by given hash or number.
      operationId: getBlock
      parameters:
        - $ref: "#/components/parameters/hashOrNumberPathParam"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Block"
        400:
          $ref: "#/components/responses/InvalidParameterError"
        404:
          $ref: "#/components/responses/NotFoundError"
      tags:
        - blocks
  /api/v1/contracts/call:
    post:
      summary: Invoke a smart contract
      description:
        Returns a result from EVM execution such as cost-free execution of read-only smart contract queries, gas estimation, and transient simulation of read-write operations. If the `estimate` field is set to true gas estimation is executed.
        This API can process calls against the `latest` block or specific historical blocks when a hexadecimal or decimal block number is provided in the `block` field.
      operationId: contractCall
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContractCallRequest"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractCallResponse"
        400:
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        404:
          description: Not found error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        415:
          description: Unsupported media type error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        429:
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        500:
          description: Generic error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        501:
          description: Not implemented error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      tags:
        - contracts
  /api/v1/contracts/simulate:
    post:
      summary: Simulate a bundle of smart contract calls
      description:
        Backs `eth_simulateV1`. Each `block_state_calls` entry executes as its own virtual block against the
        single anchor block selected by `block` - calls within an entry are cumulative with each other, and real
        execution state resets back to the anchor baseline between entries. `state_overrides` are the exception -
        they persist and accumulate across entries (later entries win per address). `transaction_index` and
        `log_index` restart from 0 at the beginning of every entry. A reverted call does not abort the rest of
        the batch. `state_overrides` reuses the same schema as `POST /api/v1/contracts/call` and is subject to
        the same feature flag.
      operationId: contractsSimulate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SimulateRequest"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimulateResponse"
        400:
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        415:
          description: Unsupported media type error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        429:
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        500:
          description: Generic error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      tags:
        - contracts
  /api/v1/contracts:
    get:
      summary: List contract entities on network
      description: Returns a list of all contract entity items on the network.
      operationId: getContracts
      parameters:
        - $ref: "#/components/parameters/contractIdQueryParam"
        - $ref: "#/components/parameters/limitQueryParam"
        - $ref: "#/components/parameters/orderQueryParamDesc"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractsResponse"
        400:
          $ref: "#/components/responses/InvalidParameterError"
      tags:
        - contracts
...
info:
  title: Mirror Node REST API
  version: 0.164.0-SNAPSHOT
  license:
    name: Apache-2.0
    url: "https://www.apache.org/licenses/LICENSE-2.0.html"
  description: "The REST API offers the ability to query transactions and entity information from a mirror node.\n\nBase url: [/api/v1](/api/v1)\n\nOpenAPI Spec: [/api/v1/docs/openapi.yml](/api/v1/docs/openapi.yml)"
  contact:
    name: Mirror Node Team
    email: mirrornode@hedera.com
    url: "https://github.com/hiero-ledger/hiero-mirror-node"
...
servers:
  - description: The current REST API server
    url: ""
  - description: The production REST API servers
    url: "{scheme}://{network}.mirrornode.hedera.com"
    variables:
      scheme:
        default: https
        description: The URI scheme
        enum: [http, https]
      network:
        default: testnet
        description: The Hedera network in use
        enum: [mainnet-public, mainnet, previewnet, testnet]
components:
  schemas:
...
    ContractsResponse:
      type: object
      properties:
        contracts:
          $ref: "#/components/schemas/Contracts"
        links:
          $ref: "#/components/schemas/Links"
...
    Contract:
      type: object
      properties:
        admin_key:
          $ref: "#/components/schemas/Key"
        auto_renew_account:
          $ref: "#/components/schemas/EntityId"
        auto_renew_period:
          example: 7776000
          format: int64
          type: [integer, "null"]
        contract_id:
          $ref: "#/components/schemas/EntityId"
        created_timestamp:
          $ref: "#/components/schemas/TimestampNullable"
        deleted:
          type: boolean
          example: false
        evm_address:
          $ref: "#/components/schemas/EvmAddress"
        expiration_timestamp:
          $ref: "#/components/schemas/TimestampNullable"
        file_id:
          $ref: "#/components/schemas/EntityId"
        max_automatic_token_associations:
          type: [integer, "null"]
          format: int32
        memo:
          example: contract memo
          type: string
        nonce:
          description: The nonce of the contract
          format: int64
          type: [integer, "null"]
          example: 1
        obtainer_id:
          $ref: "#/components/schemas/EntityId"
        permanent_removal:
          type: [boolean, "null"]
        proxy_account_id:
          $ref: "#/components/schemas/EntityId"
        timestamp:
          $ref: "#/components/schemas/TimestampRange"
    Contracts:
      type: array
      items:
        $ref: "#/components/schemas/Contract"
...
    EntityId:
      type: [string, "null"]
      description: "Network entity ID in the format of `shard.realm.num`"
      pattern: '^\d{1,10}\.\d{1,10}\.\d{1,10}$'
      example: "0.0.2"
...
    Error:
      type: object
      properties:
        _status:
          type: object
          properties:
            messages:
              type: array
              items:
                type: object
                properties:
                  data:
                    description: Error message in hexadecimal
                    example: "0x3000"
                    format: binary
                    pattern: "^0x[0-9a-fA-F]+$"
                    type: [string, "null"]
                  detail:
                    description: Detailed error message
                    example: Generic detailed error message
                    type: [string, "null"]
                  message:
                    description: Error message
                    example: Generic error message
                    type: string
    ContractCallRequest:
      type: object
      properties:
        block:
          description: Hexadecimal block number or the string "latest", "pending", "earliest". Defaults to "latest".
          example: "latest"
          pattern: "^((0x)?[0-9a-fA-F]+|(earliest|pending|latest))$"
          type: [string, "null"]
        data:
          description: Hexadecimal method signature and encoded parameters. Up to 131072 bytes as at most 262146 hexadecimal digits including optional leading 0x.
          example: "0x47f1aae7"
          format: binary
          maxLength: 262146
          pattern: "^(0x)?[0-9a-fA-F]+$"
          type: [string, "null"]
        estimate:
          description: Whether gas estimation is called. Defaults to false.
          example: true
          type: [boolean, "null"]
        from:
          description: The 20-byte hexadecimal EVM address the transaction is sent from.
          example: "00000000000000000000000000000000000004e2"
          format: binary
          maxLength: 42
          minLength: 40
          pattern: "^(0x)?[A-Fa-f0-9]{40}$"
          type: [string, "null"]
        gas:
          description: Gas provided for the transaction execution. Defaults to 15000000.
          example: 15000000
          format: int64
          minimum: 0
          type: [integer, "null"]
        gasPrice:
          description: Gas price used for each paid gas.
          example: 100000000
          format: int64
          minimum: 0
          type: [integer, "null"]
        to:
          description: The 20-byte hexadecimal EVM address the transaction is directed to.
          example: "0xd9d0c5c0ff85758bdf05a7636f8036d4d065f5b6"
          format: binary
          maxLength: 42
          minLength: 40
          pattern: "^(0x)?[A-Fa-f0-9]{40}$"
          type: string
        value:
          description: Value sent with this transaction. Defaults to 0.
          example: 0
          format: int64
          minimum: 0
          type: [integer, "null"]
      required:
        - to
    ContractCallResponse:
      type: object
      properties:
        result:
          description: Result in hexadecimal from executed contract call.
          example: "0x0000000000006d8d"
          format: binary
          pattern: "^0x[0-9a-fA-F]+$"
          type: string
    StateOverride:
      description: Per-address state override for `POST /api/v1/contracts/call` and `POST /api/v1/contracts/simulate`. All fields are optional except `address`. `state` and `state_diff` are mutually exclusive.
      type: object
      properties:
        address:
          description: The 20-byte hexadecimal EVM address of the account to override.
          example: "0x00000000000000000000000000000000000004e4"
          format: binary
          maxLength: 42
          minLength: 40
          pattern: "^(0x)?[A-Fa-f0-9]{40}$"
          type: string
        balance:
          description: Hexadecimal balance override in tinybars.
          example: "0xFFFFFFFFFFFFFFFF"
          format: binary
          maxLength: 18
          type: [string, "null"]
        code:
          description: Hexadecimal runtime bytecode override.
          example: "0x602a60005260206000f3"
          format: binary
          type: [string, "null"]
        nonce:
          description: Hexadecimal Ethereum nonce override.
          example: "0x1"
          format: binary
          maxLength: 18
          type: [string, "null"]
        state:
          description: Full storage replacement. All existing storage for this address is discarded; only these slots exist. Mutually exclusive with `state_diff`.
          items:
            $ref: "#/components/schemas/StorageEntry"
          maxItems: 100
          type: [array, "null"]
        state_diff:
          description: Storage patch. Only the listed slots are overridden; all other slots fall through to the underlying state. Mutually exclusive with `state`.
          items:
            $ref: "#/components/schemas/StorageEntry"
          maxItems: 100
          type: [array, "null"]
      required:
        - address
    StorageEntry:
      description: A single storage slot key-value pair used within a StateOverride.
      type: object
      properties:
        key:
          description: 32-byte hexadecimal storage slot identifier.
          example: "0x0000000000000000000000000000000000000000000000000000000000000002"
          format: binary
          maxLength: 66
          minLength: 64
          type: string
        value:
          description: 32-byte hexadecimal storage slot value.
          example: "0x000000000000000000000000000000000000000000000000000000000000270f"
          format: binary
          maxLength: 66
          minLength: 64
          type: string
      required:
        - key
        - value
    SimulateCall:
      type: object
      properties:
        data:
          description: Hexadecimal method signature and encoded parameters.
          example: "0x47f1aae7"
          format: binary
          pattern: "^(0x)?[0-9a-fA-F]+$"
          type: [string, "null"]
        from:
          description: The 20-byte hexadecimal EVM address the call is sent from.
          example: "0xd8dA6BF26964aF9D7eEd9e03E53415D37aA96045"
          format: binary
          maxLength: 42
          minLength: 40
          pattern: "^(0x)?[A-Fa-f0-9]{40}$"
          type: [string, "null"]
        gas:
          description: Gas provided for the call execution, as a number or a hexadecimal/decimal string. Defaults to 15000000.
          example: 15000000
          minimum: 21000
          type: [integer, string, "null"]
        gasPrice:
          description: Gas price used for each paid gas, as a number or a hexadecimal/decimal string.
          example: 0
          minimum: 0
          type: [integer, string, "null"]
        max_fee_per_gas:
          description: Accepted for compatibility with the HIP's example request, but unused - Mirror Node has no EIP-1559 fee market.
          example: "0xf"
          format: binary
          type: [string, "null"]
        to:
          description: The 20-byte hexadecimal EVM address the call is directed to.
          example: "0x014d023e954bAae7F21E56ed8a5d81b12902684D"
          format: binary
          maxLength: 42
          minLength: 40
          pattern: "^(0x)?[A-Fa-f0-9]{40}$"
          type: [string, "null"]
        value:
          description: Value sent with this call, as a number or a hexadecimal/decimal string (e.g. "0x1"). Defaults to 0.
          example: 1
          minimum: 0
          type: [integer, string, "null"]
    SimulateBlockStateCall:
      type: object
      properties:
        calls:
          items:
            $ref: "#/components/schemas/SimulateCall"
          type: [array, "null"]
        state_overrides:
          description: State overrides taking effect when this entry begins, persisting for the rest of the request (later entries win per address).
          items:
            $ref: "#/components/schemas/StateOverride"
          type: [array, "null"]
    SimulateRequest:
      type: object
      properties:
        block:
          description: Hexadecimal block number or the string "latest", "pending", "earliest". Defaults to "latest". Anchors the whole request - all block_state_calls entries execute against this one block.
          example: "latest"
          pattern: "^((0x)?[0-9a-fA-F]+|(earliest|pending|latest))$"
          type: [string, "null"]
        block_state_calls:
          description: Each entry executes as its own virtual block against the anchor block - cumulative within an entry, state reset between entries. Limited to 16 total calls across all entries.
          items:
            $ref: "#/components/schemas/SimulateBlockStateCall"
          type: [array, "null"]
        trace_transfers:
          description: If true, synthesizes ERC20-style Transfer log entries for value transfers captured during execution. Defaults to false.
          example: true
          type: [boolean, "null"]
    SimulateLog:
      type: object
      properties:
        address:
          description: The 20-byte hexadecimal EVM address that emitted this log.
          format: binary
          type: string
        block_hash:
          format: binary
          type: string
        block_number:
          format: binary
          type: string
        data:
          format: binary
          type: string
        log_index:
          format: binary
          type: string
        removed:
          type: boolean
        topics:
          items:
            format: binary
            type: string
          type: array
        transaction_hash:
          description: Synthetic identifier - simulated calls are never mined, so there is no real transaction hash.
          format: binary
          type: string
        transaction_index:
          format: binary
          type: string
    SimulateCallResult:
      type: object
      properties:
        gas_used:
          description: Hexadecimal gas consumed by this call.
          example: "0x5208"
          format: binary
          type: string
        logs:
          items:
            $ref: "#/components/schemas/SimulateLog"
          type: array
        return_data:
          description: Hexadecimal return value of this call.
          example: "0x"
          format: binary
          type: string
        status:
          description: '"0x1" on success, "0x0" on revert. A reverted call does not abort the rest of the batch.'
          example: "0x1"
          type: string
    SimulateResponse:
      type: object
      properties:
        result:
          description: One inner array per block_state_calls entry, in order; each inner array maps 1:1, in order, to that entry's calls.
          items:
            items:
              $ref: "#/components/schemas/SimulateCallResult"
            type: array
          type: array
...
    EvmAddress:
      type: string
      description: A network entity encoded as an EVM address in hex.
      format: binary
      minLength: 40
      maxLength: 42
      pattern: "^(0x)?[A-Fa-f0-9]{40}$"
      example: "0000000000000000000000000000000000001f41"
...
    Key:
      description: The public key which controls access to various network entities.
      type: [object, "null"]
      properties:
        _type:
          type: string
          enum: [ECDSA_SECP256K1, ED25519, ProtobufEncoded]
          example: ProtobufEncoded
        key:
          type: string
          example: "15706b229b3ba33d4a5a41ff54ce1cfe0a3d308672a33ff382f81583e02bd743"
    Links:
      type: object
      properties:
        next:
          example: null
          type: [string, "null"]
...
    Block:
      type: object
      properties:
        count:
          type: integer
          minimum: 0
        gas_used:
          type: [integer, "null"]
          minimum: 0
          format: int64
        hapi_version:
          type: [string, "null"]
        hash:
          type: string
        logs_bloom:
          description: A hex encoded 256-byte array with 0x prefix
          type: [string, "null"]
          pattern: ^0x[0-9a-fA-F]{512}$
        name:
          type: string
        number:
          type: integer
          minimum: 0
        previous_hash:
          type: string
        size:
          type: [integer, "null"]
        timestamp:
          $ref: "#/components/schemas/TimestampRange"
      example:
        count: 3
        gas_used: 300000
        hapi_version: "0.11.0"
        hash: "0x3c08bbbee74d287b1dcd3f0ca6d1d2cb92c90883c4acf9747de9f3f3162ad25b999fc7e86699f60f2a3fb3ed9a646c6b"
        logs_bloom: "0x00000020002000001000000000000000000000000000000000000000000010000000000004000000000000000000000000108000000000000000000080000000000004000000000000000000000000880000000000000000000101000000000000000000000000000000000000008000000000000400000080000000000001000000000000000000000000000000000000000000002000000000100000100000200000040000100000001000000000000000000000000000000001001000004000000000000000000001000000000000000000100000000000100000000000000000000000000000000000000000000000080000100800000000000000120080"
        name: "2022-05-03T06_46_26.060890949Z.rcd"
        number: 77
        previous_hash: "0xf7d6481f659c866c35391ee230c374f163642ebf13a5e604e04a95a9ca48a298dc2dfa10f51bcbaab8ae23bc6d662a0b"
        size: 8192
        timestamp:
          from: "1651560386.060890949"
          to: "1651560386.661997287"
...
    Timestamp:
      description: A Unix timestamp in seconds.nanoseconds format
      type: string
      example: "1586567700.453054000"
      pattern: '^\d{1,10}(\.\d{1,9})?$'
    TimestampNullable:
      description: A Unix timestamp in seconds.nanoseconds format
      type: [string, "null"]
      example: "1586567700.453054000"
      pattern: '^\d{1,10}(\.\d{1,9})?$'
    TimestampRange:
      type: object
      description: A timestamp range an entity is valid for
      properties:
        from:
          allOf:
            - $ref: "#/components/schemas/Timestamp"
            - description: The inclusive from timestamp in seconds
        to:
          allOf:
            - $ref: "#/components/schemas/TimestampNullable"
            - description: The exclusive to timestamp in seconds
...
  responses:
    NotFoundError:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            _status:
              messages:
                - message: Not found
...
    InvalidParameterError:
      description: Invalid parameter
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            _status:
              messages:
                - message: "Invalid parameter: account.id"
                - message: Invalid Transaction id. Please use \shard.realm.num-sss-nnn\ format where sss are seconds and nnn are nanoseconds
...
  parameters:
...
    contractIdQueryParam:
      name: contract.id
      description: The ID of the smart contract
      in: query
      examples:
        noValue:
          summary: --
          value: ""
        entityNumNoOperator:
          summary: Example of entityNum equals with no operator
          value: 100
        idNoOperator:
          summary: Example of id equals with no operator
          value: 0.0.100
        evmAddressNoOperator:
          summary: A hex encoded EVM address with no operator.
          value: 0.0.1d8bfdc5d46dc4f61d6b6115972536ebe6a8854c
        entityNumEqOperator:
          summary: Example of entityNum equals operator
          value: eq:200
        idEqOperator:
          summary: Example of id equals operator
          value: eq:0.0.200
        entityNumNeOperator:
          summary: Example of entityNum not equals operator
          value: ne:300
        idNeOperator:
          summary: Example of id not equals operator
          value: ne:0.0.300
        entityNumGtOperator:
          summary: Example of entityNum greater than operator
          value: gt:400
        idGtOperator:
          summary: Example of id greater than operator
          value: gt:0.0.400
        entityNumGteOperator:
          summary: Example of entityNum greater than or equals operator
          value: gte:500
        idGteOperator:
          summary: Example of id greater than or equals operator
          value: gte:0.0.500
        entityNumLtOperator:
          summary: Example of entityNum less than operator
          value: lt:600
        idLtOperator:
          summary: Example of id less than operator
          value: lt:0.0.600
        entityNumLteOperator:
          summary: Example of entityNum less than or equals operator
          value: lte:700
        idLteOperator:
          summary: Example of id less than or equals operator
          value: lte:0.0.700
      schema:
        pattern: ^((gte?|lte?|eq|ne)\:)?(\d{1,10}\.\d{1,10}\.)?\d{1,10}|(\d{1,10}\.){0,2}[A-Fa-f0-9]{40}$
        type: string
...
    hashOrNumberPathParam:
      name: hashOrNumber
      in: path
      required: true
      description: Accepts both eth and hedera hash format or block number
      schema:
        pattern: ^(\d{1,10}|(0x)?([A-Fa-f0-9]{64}|[A-Fa-f0-9]{96}))$
        type: string
...
    limitQueryParam:
      name: limit
      in: query
      description: The maximum number of items to return
      example: 2
      schema:
        format: int32
        type: integer
        default: 25
        minimum: 1
        maximum: 100
...
    orderQueryParamDesc:
      name: order
      in: query
      description: The order in which items are listed
      example: asc
      schema:
        enum: [asc, desc]
        default: desc
...